Add dry-run mode.

+++ slurm/compress.py
@@ -22,9 +22,16 @@ if __name__ == '__main__':
             files_to_compress.append(file)
     print(f"Compressing {len(files_to_compress)} directories...")
     time = datetime.now()
-    os.system(f"tar -czvf {parent_directory}/compressed_{time}.tar.gz {' '.join(str(files_to_compress))}")
+    if "--dry-run" not in args:
+        os.system(f"tar -czvf {parent_directory}/compressed_{time}.tar.gz {' '.join(str(files_to_compress))}")
+    else:
+        print(f"tar -czvf {parent_directory}/compressed_{time}.tar.gz {' '.join(str(files_to_compress))}")
+        print(f"Dry run, not compressing {len(files_to_compress)} runs.")
     print(f"Done compressing {len(files_to_compress)} directories.")
     if "-d" in args or "--delete" in args:
         print("Cleaning runs...")
         for file in files_to_compress:
-            file.unlink()
+            if "--dry-run" not in args:
+                file.unlink()
+            else:
+                print("[dry-run mode] would deleting", file)
